Fix handling of empty responses
Previously, the application would crash upon receiving an empty response,
because it would attempt to assign an empty arraylist to the array adapter.
Now, the REST API returns an empty JSON object if the query returned no results,
and the application shows a toast to the user stating "No results found"
instead of starting the results activity.